+++ dwds/web/client.dart
@@ -33,7 +33,7 @@ Future<void> main() {
     // Test apps may already have this set.
     dartAppInstanceId ??= Uuid().v1();
 
-    var client = SseClient(r'/$sseHandler');
+    var client = SseClient(_fixProtocol('$dartUriBase/\$sseHandler'));
     // Ensure the SSE connection is established before proceeding.
     // Note that `onOpen` is a broadcast stream so we must listen for this
     // immediately.
@@ -79,7 +79,6 @@ Future<void> main() {
           window.alert('DevTools failed to open with: ${event.error}');
         }
       } else if (event is RunRequest) {
-        client.sink.add(jsonEncode(serializers.serialize(RunResponse())));
         runMain();
       } else if (event is ErrorResponse) {
         window.console.error('Error from backend:\n\nError: ${event.error}\n\n'
@@ -133,11 +132,30 @@ $stackTrace
   });
 }
 
+/// Returns [url] modified if necessary so that, if the current page is served
+/// over `https`, then the URL is converted to `https`. Localhost is treated
+/// as a special case and not modified.
+String _fixProtocol(String url) {
+  if (window.location.protocol == 'https:' && !url.startsWith('https://')) {
+    // Chrome seems to allow mixed content from localhost.
+    if (url.startsWith('http://localhost')) {
+      return url;
+    } else {
+      return url.replaceFirst('http://', 'https://');
+    }
+  } else {
+    return url;
+  }
+}
+
 @JS(r'$dartAppId')
 external String get dartAppId;
 
 @JS(r'$dartAppInstanceId')
 external String get dartAppInstanceId;
+
+@JS(r'$dartUriBase')
+external String get dartUriBase;
 
 @JS(r'$dartAppInstanceId')
 external set dartAppInstanceId(String id);
@@ -157,8 +175,15 @@ external set launchDevToolsJs(void Function() cb);
 @JS(r'$dartReloadConfiguration')
 external String get reloadConfiguration;
 
-@JS(r'$dartRunMain')
-external void Function() get runMain;
+/// Runs `window.$dartRunMain()` by injecting a script tag.
+///
+/// We do this so that we don't see user exceptions bubble up in our own error
+/// handling zone.
+void runMain() {
+  var scriptElement = ScriptElement()..innerHtml = r'window.$dartRunMain();';
+  document.body.append(scriptElement);
+  Future.microtask(scriptElement.remove);
+}
 
 bool get _isChrome =>
     window.navigator.userAgent.contains('Chrome') &&